requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.16
django-allauth==64.2.1
django-apscheduler==0.6.2
django-filter==24.3
djangorestframework==3.15.2
flake8==7.1.1
idna==3.10
importlib_metadata==8.5.0
isort==5.13.2
kombu==5.4.2
Markdown==3.7
mccabe==0.7.0
prompt_toolkit==3.0.47
pycodestyle==2.12.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.0.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
zipp==3.20.2
